requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
PyJWT>=2.11.0
bcrypt>=4.0.0
cryptography>=42.0.0
anthropic>=0.84.0
openai>=2.0.0
pyotp>=2.9.0
pdfplumber>=0.11.0
python-multipart>=0.0.20
qrcode>=8.0
plaid-python>=16.0.0
apscheduler>=3.10.0
httpx>=0.28.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
setuptools>=70.0.0
